feat: Count voice state sessions for desk channels
This keeps a counter of users connected to a desk to mark the desk as hidden only when the last user disconnects.

Ideally, we should have access to the number of users connected to a voice channel, or the list of current voice sessions for a guild on the event.